requirements.txt
Flask==0.10.1
geopy==1.10.0
python-google-places==1.0.0